Avoid passing null pointers for ustm, hfx, and qfx to atm_compute_dyn_tend_work
The dummy arguments for ustm, hfx, and qfx in the atm_compute_dyn_tend_work
routine are not pointers, and so passing unassociated pointers for the actual
arguments is invalid and will trigger a runtime error when debugging options are
enabled with the GNU compilers.

The workaround adopted in this commit is to allocate the ustm, hfx, and qfx
array pointers with a trivial size if these fields are not available (due to
packages associated with PBL schemes), and to deallocate them after the call to
atm_compute_dyn_tend_work. The dummy arguments for ustm, hfx, and qfx in
atm_compute_dyn_tend_work (as well as in u_dissipation_3d and
scalar_dissipation_3d_les) are now assumed-shape arrays, as the actual arguments
may not have (nCells+1) elements.